bear/requirements.txt
# Pipecat Voice AI Framework
pipecat-ai[silero,openai,daily]>=0.0.43

# Audio handling
pyaudio>=0.2.14
numpy>=1.24.0

# Snowflake API calls
httpx>=0.27.0
python-dotenv>=1.0.0

# For Pi GPIO (optional - for button trigger)
# RPi.GPIO>=0.7.1

